Move downloader service to separate module
